+++ app/models/report.rb
@@ -11,7 +11,6 @@ class Report < ApplicationRecord
   has_many :transactions
   has_many :planned_disbursements
 
-  validates_uniqueness_of :fund, scope: :organisation
   validate :activity_must_be_a_fund
   validates :deadline, date_not_in_past: true, date_within_boundaries: true