+++ db/data_schema.rb
@@ -1,2 +1,2 @@
 # encoding: UTF-8
-DataMigrate::Data.define(version: 20200903140938)
+DataMigrate::Data.define(version: 20200907144609)