+++ db/migrate/20200907141532_add_gdi_to_activities.rb
@@ -1,0 +1,5 @@
+class AddGdiToActivities < ActiveRecord::Migration[6.0]
+  def change
+    add_column :activities, :gdi, :string
+  end
+end