+++ spec/features/staff/users_can_edit_an_activity_spec.rb
@@ -190,7 +190,7 @@ RSpec.feature "Users can edit an activity" do
           select recipient_region, from: "activity[recipient_region]"
           click_button t("form.button.activity.submit")
 
-          expect(page).to have_content t("form.label.activity.flow")
+          expect(page).to have_content t("form.label.activity.intended_beneficiaries")
           expect(page).not_to have_content activity.title
         end
 
@@ -477,6 +477,24 @@ def assert_all_edit_links_go_to_the_correct_form_step(activity:)
   click_on(t("default.link.back"))
   click_on t("tabs.activity.details")
 
+  within(".intended_beneficiaries") do
+    click_on(t("default.link.edit"))
+    expect(page).to have_current_path(
+      activity_step_path(activity, :intended_beneficiaries)
+    )
+  end
+  click_on(t("default.link.back"))
+  click_on t("tabs.activity.details")
+
+  within(".gdi") do
+    click_on(t("default.link.edit"))
+    expect(page).to have_current_path(
+      activity_step_path(activity, :gdi)
+    )
+  end
+  click_on(t("default.link.back"))
+  click_on t("tabs.activity.details")
+
   within(".flow") do
     click_on(t("default.link.edit"))
     expect(page).to have_current_path(
@@ -490,6 +508,15 @@ def assert_all_edit_links_go_to_the_correct_form_step(activity:)
     click_on(t("default.link.edit"))
     expect(page).to have_current_path(
       activity_step_path(activity, :aid_type)
+    )
+  end
+  click_on(t("default.link.back"))
+  click_on t("tabs.activity.details")
+
+  within(".oda_eligibility") do
+    click_on(t("default.link.edit"))
+    expect(page).to have_current_path(
+      activity_step_path(activity, :oda_eligibility)
     )
   end
   click_on(t("default.link.back"))