+++ spec/models/activity_spec.rb
@@ -433,10 +433,38 @@ RSpec.describe Activity, type: :model do
       end
     end
 
+    context "when requires_additional_benefitting_countries is blank when required" do
+      subject(:activity) { build(:activity, geography: :recipient_country, requires_additional_benefitting_countries: nil) }
+      it "should not be valid" do
+        expect(activity.valid?(:requires_additional_benefitting_countries_step)).to be_falsey
+      end
+    end
+
+    context "when intended_beneficiaries is blank" do
+      subject(:activity) { build(:activity, intended_beneficiaries: nil) }
+      it "should not be valid" do
+        expect(activity.valid?(:intended_beneficiaries_step)).to be_falsey
+      end
+    end
+
+    context "when gdi is blank" do
+      subject(:activity) { build(:activity, gdi: nil) }
+      it "should not be valid" do
+        expect(activity.valid?(:gdi_step)).to be_falsey
+      end
+    end
+
     context "when flow is blank" do
       subject(:activity) { build(:activity, flow: nil) }
       it "should not be valid" do
         expect(activity.valid?(:flow_step)).to be_falsey
+      end
+    end
+
+    context "when oda_eligibility is blank" do
+      subject(:activity) { build(:activity, oda_eligibility: nil) }
+      it "should not be valid" do
+        expect(activity.valid?(:oda_eligibility_step)).to be_falsey
       end
     end