+++ spec/services/ingest_iati_activities_spec.rb
@@ -112,6 +112,8 @@ RSpec.describe IngestIatiActivities do
       expect(activity.recipient_region).to eql("998")
       expect(activity.recipient_country).to be_nil
 
+      expect(activity.intended_beneficiaries).to eql(["Replace me"])
+      expect(activity.gdi).to eql("Replace me")
       expect(activity.sector).to eql("43082")
       expect(activity.flow).to eql("10")
       expect(activity.aid_type).to eql("C01")